migrate notification center code to a separate component
