+++ app/routes/testpage/testpage.module.css
@@ -1,51 +1,3 @@
 .testpage {
     background-color: green;
 }
-
-.notifications {
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 15px;
-}
-
-.notification {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: stretch;
-    width: 380px;
-    height: 90px;
-    padding: 16px;
-    font-family: var(--font-family-main);
-    background-color: var(--dark3);
-    border-radius: 16px;
-}
-
-.notification > header {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.notification > header > .header_content {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 8px;
-}
-
-.notification h2 {
-    font-size: var(--font-size-m);
-}
-
-.notification > p {
-    font-size: var(--font-size-s);
-}
-
-.close {
-    cursor: pointer;
-}
